requirements.txt
amqp==5.1.1
asgiref==3.5.2
billiard==3.6.4.0
black==22.10.0
cachetools==5.2.0
celery==5.2.7
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
-e git+https://github.com/bcdevworks/conekta-python.git@a46819a4c95d5e62b75eb85516495e21c095d1fa#egg=conekta
Django==4.1.3
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
flake8==5.0.4
google-api-core==2.10.2
google-api-python-client==2.66.0
google-auth==2.14.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.7.1
google-cloud-core==2.3.2
google-cloud-secret-manager==2.12.6
google-cloud-storage==2.6.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
googlemaps==4.7.0
grpc-google-iam-v1==0.12.4
grpcio==1.50.0
grpcio-status==1.50.0
httplib2==0.21.0
idna==3.4
isort==5.10.1
kombu==5.2.4
mccabe==0.7.0
mypy-extensions==0.4.3
nose==1.3.7
oauthlib==3.2.2
pathspec==0.10.2
platformdirs==2.5.4
prompt-toolkit==3.0.32
proto-plus==1.22.1
protobuf==4.21.9
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.6.0
pyparsing==3.0.9
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
simplejson==3.18.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
